Attempted to flesh out the account edit but I really don't understand both A: Python and B: The pylons syntactogasm for making joins work. So the shell of the edit function is there and the template will work once the variable names are set properly in the values. I've also redacted the * for required fields and replaced it with a snazzy javascript thing-a-ma-bob that indicates fields that are required when you focus into them (or out of them as the case might be in the event that someone doesn't like it.
This method is reusable on any form by applying a "req" selector to the div.f selector stack. The javascript will determine what needs the behavior applied and act appropriately. There will be places where fields are required that are kind of a given; in which case we don't need to concern ourselves with showing the adornment.

Keep on truckin'